Move find_imports function
